% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metKernel.R
\name{metKernel}
\alias{metKernel}
\title{metKernel}
\usage{
metKernel(data450, data850, chrom, path, sigma = NULL)
}
\arguments{
\item{data450}{An HM450 data set with CpG sites as columns}

\item{data850}{An EPIC probe data set with CpG sites as columns}

\item{chrom}{Indicator of chromosome (e.g. 22 or 'X')}

\item{path}{Output directory for imputed EPIC probe set and summary statistics}

\item{sigma}{Optional, if present fixes tuning parameter to value}
}
\description{
Use radial basis function (RBF) kernel method to impute level of methylation for EPIC platform CpG sites from an HM450 probe set
}
\examples{
metKernel(HM450, EPIC, 22, '/home/project/results')
metKernel(HM450, EPIC, 'X', '/home/project/results', sigma = 0.001)
}
