% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metRidge.R
\name{metRidge}
\alias{metRidge}
\title{metRidge}
\usage{
metRidge(data450, data850, chrom, path)
}
\arguments{
\item{data450}{An HM450 probe data set with CpG sites as columns}

\item{data850}{An EPIC probe data set with CpG sites as columns}

\item{chrom}{Indicator of chromosome (e.g. 22 or 'X')}

\item{path}{Output directory for imputed EPIC probe set and summary statistics}
}
\description{
Use penalized ridge regression to impute level of methylation for EPIC platform CpG sites from an HM450 probe set
}
\examples{
metRidge(HM450, EPIC, 22, '/home/project/results')
metRidge(HM450, EPIC, 'X', '/home/project/results')
}
