% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression}
\usage{
ridge(data450, data850, chrom, path)
}
\arguments{
\item{data450}{An HM450 probe data set with CpG sites as columns}

\item{data850}{An EPIC probe data set with CpG sites as columns}

\item{chrom}{Indicator of chromosome (e.g. 22 or 'X')}

\item{path}{Output directory for imputed EPIC probe set and summary statistics}
}
\description{
Use penalized ridge regression to impute EPIC platform CpG sites from an HM450 probe set
}
\examples{
ridge(HM450, EPIC, 22, '/home/project/results')
ridge(HM450, EPIC, 'X', '/home/project/results')
}
